refactor: change gift package name to product
