+++ greengift/src/main/java/com/apiserver/greengift/festival/FestivalWriteService.java
@@ -5,13 +5,13 @@ import com.apiserver.greengift._core.errors.exception.BadRequestException;
 import com.apiserver.greengift._core.errors.exception.NotFoundException;
 import com.apiserver.greengift._core.errors.exception.UnauthorizedException;
 import com.apiserver.greengift.festival.constant.FestivalStatus;
-import com.apiserver.greengift.gift.constant.ProductCategory;
-import com.apiserver.greengift.gift.Product;
-import com.apiserver.greengift.gift.ProductJPARepository;
+import com.apiserver.greengift.product.constant.ProductCategory;
+import com.apiserver.greengift.product.Product;
+import com.apiserver.greengift.product.ProductJPARepository;
 import com.apiserver.greengift.festival.user_festival.UserFestival;
 import com.apiserver.greengift.festival.user_festival.UserFestivalJPARepository;
-import com.apiserver.greengift.gift.user_product.UserProduct;
-import com.apiserver.greengift.gift.user_product.UserProductJPARepository;
+import com.apiserver.greengift.product.user_product.UserProduct;
+import com.apiserver.greengift.product.user_product.UserProductJPARepository;
 import com.apiserver.greengift.user.User;
 import com.apiserver.greengift.user.festival_manager.FestivalManager;
 import com.apiserver.greengift.user.festival_manager.FestivalManagerJPARepository;
