+++ greengift/src/main/java/com/apiserver/greengift/product/ProductJPARepository.java
@@ -1,4 +1,4 @@
-package com.apiserver.greengift.gift;
+package com.apiserver.greengift.product;
 
 import com.apiserver.greengift.festival.Festival;
 import org.springframework.data.jpa.repository.JpaRepository;
