+++ greengift/src/main/java/com/apiserver/greengift/product/ProductResponse.java
@@ -1,0 +1,5 @@
+package com.apiserver.greengift.product;
+
+public class ProductResponse {
+
+}
