+++ greengift/src/main/java/com/apiserver/greengift/product/ProductService.java
@@ -1,0 +1,4 @@
+package com.apiserver.greengift.product;
+
+public class ProductService {
+}
