+++ greengift/src/main/java/com/apiserver/greengift/product/user_product/UserProductJPARepository.java
@@ -1,4 +1,4 @@
-package com.apiserver.greengift.gift.user_product;
+package com.apiserver.greengift.product.user_product;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
